pyproject.toml
[tool.poetry]
name = "drfcw"
version = "0.1.0"
description = ""
authors = ["Dmitriy Kucheriaviy <dmitask@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.4.0"
django-filter = "^24.3"
python-telegram-bot = "^21.10"
celery = "^5.4.0"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.6.0"
drf-spectacular = {extras = ["redoc", "swagger-ui"], version = "^0.28.0"}
flake8 = "^7.1.1"
asgiref = "^3.8.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-django = "^4.9.0"
pytest-cov = "^6.0.0"
pytest-celery = "^1.1.3"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
